chore(ci): update sweeper registration for Metal resources
