+++ GNUmakefile
@@ -3,7 +3,7 @@ TEST				?=$$(go list ./... |grep -v 'vendor'|grep -v 'tests')
 INSTALL_DIR			=~/.terraform.d/plugins
 BINARY				=terraform-provider-equinix
 SWEEP				?=all #Flag required to define the regions that the sweeper is to be ran in
-SWEEP_DIR			?=./equinix
+SWEEP_DIR			?=$$(go list ./... | grep 'internal/sweep\|equinix/equinix')
 SWEEP_ARGS			=-timeout 60m
 ACCTEST_TIMEOUT		?= 180m
 ACCTEST_PARALLELISM ?= 8
