+++ equinix/resource_metal_user_api_key_acc_test.go
@@ -1,48 +1,6 @@
 package equinix
 
-import (
-	"fmt"
-	"log"
-	"net/http"
-
-	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
-	// "github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-)
-
-func init() {
-	resource.AddTestSweepers("equinix_metal_user_api_key", &resource.Sweeper{
-		Name: "equinix_metal_user_api_key",
-		F:    testSweepUserAPIKeys,
-	})
-}
-
-// Will remove all User API keys with Description starting with "tfacc-"
-func testSweepUserAPIKeys(region string) error {
-	log.Printf("[DEBUG] Sweeping user_api keys")
-	config, err := sharedConfigForRegion(region)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting configuration for sweeping user_api keys: %s", err)
-	}
-	metal := config.NewMetalClient()
-	userApiKeys, _, err := metal.APIKeys.UserList(nil)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting list for sweeping user_api keys: %s", err)
-	}
-	ids := []string{}
-	for _, k := range userApiKeys {
-		if isSweepableTestResource(k.Description) {
-			ids = append(ids, k.ID)
-		}
-	}
-	for _, id := range ids {
-		log.Printf("Removing user api key %s", id)
-		resp, err := metal.APIKeys.Delete(id)
-		if err != nil && resp.StatusCode != http.StatusNotFound {
-			return fmt.Errorf("Error deleting user_api key %s", err)
-		}
-	}
-	return nil
-}
+// "github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
 // Commented out because it lists existing user API keys in debug log
