+++ equinix/resource_metal_virtual_circuit.go
@@ -2,6 +2,7 @@ package equinix
 
 import (
 	"context"
+	"fmt"
 	"log"
 	"reflect"
 	"regexp"
@@ -18,6 +19,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/packethost/packngo"
 )
 
@@ -358,5 +360,53 @@ func resourceMetalVirtualCircuitDelete(ctx context.Context, d *schema.ResourceDa
 		return diag.Errorf("Error deleting virtual circuit %s: %s", d.Id(), err)
 	}
 	d.SetId("")
+	return nil
+}
+
+func addMetalVirtualCircuitSweeper() {
+	resource.AddTestSweepers("equinix_metal_virtual_circuit", &resource.Sweeper{
+		Name:         "equinix_metal_virtual_circuit",
+		Dependencies: []string{},
+		F:            testSweepVirtualCircuits,
+	})
+}
+
+func testSweepVirtualCircuits(region string) error {
+	log.Printf("[DEBUG] Sweeping VirtualCircuits")
+	config, err := sharedConfigForRegion(region)
+	if err != nil {
+		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting configuration for sweeping VirtualCircuits: %s", err)
+	}
+	metal := config.NewMetalClient()
+	orgList, _, err := metal.Organizations.List(nil)
+	if err != nil {
+		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting organization list for sweeping VirtualCircuits: %s", err)
+	}
+	vcs := map[string]*packngo.VirtualCircuit{}
+	for _, org := range orgList {
+		conns, _, err := metal.Connections.OrganizationList(org.ID, &packngo.GetOptions{Includes: []string{"ports"}})
+		if err != nil {
+			return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting connections list for sweeping VirtualCircuits: %s", err)
+		}
+		for _, conn := range conns {
+			if conn.Type != packngo.ConnectionShared {
+				for _, port := range conn.Ports {
+					for _, vc := range port.VirtualCircuits {
+						if isSweepableTestResource(vc.Name) {
+							vcs[vc.ID] = &vc
+						}
+					}
+				}
+			}
+		}
+	}
+	for _, vc := range vcs {
+		log.Printf("[INFO][SWEEPER_LOG] Deleting VirtualCircuit: %s", vc.Name)
+		_, err := metal.VirtualCircuits.Delete(vc.ID)
+		if err != nil {
+			return fmt.Errorf("[INFO][SWEEPER_LOG] Error deleting VirtualCircuit: %s", err)
+		}
+	}
+
 	return nil
 }
