+++ equinix/resource_metal_vlan_acc_test.go
@@ -2,7 +2,6 @@ package equinix
 
 import (
 	"fmt"
-	"log"
 	"testing"
 
 	"github.com/equinix/terraform-provider-equinix/internal/config"
@@ -12,55 +11,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 	"github.com/packethost/packngo"
 )
-
-func init() {
-	resource.AddTestSweepers("equinix_metal_vlan", &resource.Sweeper{
-		Name:         "equinix_metal_vlan",
-		Dependencies: []string{"equinix_metal_virtual_circuit", "equinix_metal_vrf", "equinix_metal_device"},
-		F:            testSweepVlans,
-	})
-}
-
-func testSweepVlans(region string) error {
-	log.Printf("[DEBUG] Sweeping vlans")
-	config, err := sharedConfigForRegion(region)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting configuration for sweeping vlans: %s", err)
-	}
-	metal := config.NewMetalClient()
-	ps, _, err := metal.Projects.List(nil)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting project list for sweeping vlans: %s", err)
-	}
-	pids := []string{}
-	for _, p := range ps {
-		if isSweepableTestResource(p.Name) {
-			pids = append(pids, p.ID)
-		}
-	}
-	dids := []string{}
-	for _, pid := range pids {
-		ds, _, err := metal.ProjectVirtualNetworks.List(pid, nil)
-		if err != nil {
-			log.Printf("Error listing vlans to sweep: %s", err)
-			continue
-		}
-		for _, d := range ds.VirtualNetworks {
-			if isSweepableTestResource(d.Description) {
-				dids = append(dids, d.ID)
-			}
-		}
-	}
-
-	for _, did := range dids {
-		log.Printf("Removing vlan %s", did)
-		_, err := metal.ProjectVirtualNetworks.Delete(did)
-		if err != nil {
-			return fmt.Errorf("Error deleting vlan %s", err)
-		}
-	}
-	return nil
-}
 
 func testAccCheckMetalVlanConfig_metro(projSuffix, metro, desc string) string {
 	return fmt.Sprintf(`
