+++ internal/acceptance/acceptance.go
@@ -1,15 +1,12 @@
 package acceptance
 
 import (
-	"fmt"
 	"os"
-	"strconv"
-	"strings"
 	"testing"
-	"time"
 
 	"github.com/equinix/terraform-provider-equinix/equinix"
 	"github.com/equinix/terraform-provider-equinix/internal/config"
+	"github.com/equinix/terraform-provider-equinix/internal/env"
 	"github.com/equinix/terraform-provider-equinix/internal/provider"
 	"github.com/equinix/terraform-provider-equinix/version"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
@@ -17,8 +14,6 @@ import (
 )
 
 const (
-	// duplicated from equinix_sweeoer_test.go
-	tstResourcePrefix = "tfacc"
 	missingMetalToken = "To run acceptance tests of Equinix Metal Resources, you must set %s"
 )
 
@@ -45,15 +40,15 @@ func init() {
 func TestAccPreCheck(t *testing.T) {
 	var err error
 
-	if _, err = GetFromEnv(config.ClientTokenEnvVar); err != nil {
-		_, err = GetFromEnv(config.ClientIDEnvVar)
+	if _, err = env.Get(config.ClientTokenEnvVar); err != nil {
+		_, err = env.Get(config.ClientIDEnvVar)
 		if err == nil {
-			_, err = GetFromEnv(config.ClientSecretEnvVar)
+			_, err = env.Get(config.ClientSecretEnvVar)
 		}
 	}
 
 	if err == nil {
-		_, err = GetFromEnv(config.MetalAuthTokenEnvVar)
+		_, err = env.Get(config.MetalAuthTokenEnvVar)
 	}
 
 	if err != nil {
@@ -66,28 +61,4 @@ func TestAccPreCheckMetal(t *testing.T) {
 	if os.Getenv(config.MetalAuthTokenEnvVar) == "" {
 		t.Fatalf(missingMetalToken, config.MetalAuthTokenEnvVar)
 	}
-}
-
-func IsSweepableTestResource(namePrefix string) bool {
-	return strings.HasPrefix(namePrefix, tstResourcePrefix)
-}
-
-func GetConfigForNonStandardMetalTest() (*config.Config, error) {
-	endpoint := GetFromEnvDefault(config.EndpointEnvVar, config.DefaultBaseURL)
-	clientTimeout := GetFromEnvDefault(config.ClientTimeoutEnvVar, strconv.Itoa(config.DefaultTimeout))
-	clientTimeoutInt, err := strconv.Atoi(clientTimeout)
-	if err != nil {
-		return nil, fmt.Errorf("cannot convert value of '%s' env variable to int", config.ClientTimeoutEnvVar)
-	}
-	metalAuthToken := GetFromEnvDefault(config.MetalAuthTokenEnvVar, "")
-
-	if metalAuthToken == "" {
-		return nil, fmt.Errorf(missingMetalToken, config.MetalAuthTokenEnvVar)
-	}
-
-	return &config.Config{
-		AuthToken:      metalAuthToken,
-		BaseURL:        endpoint,
-		RequestTimeout: time.Duration(clientTimeoutInt) * time.Second,
-	}, nil
 }
