+++ internal/resources/metal/vrf/resource_test.go
@@ -3,7 +3,6 @@ package vrf_test
 import (
 	"context"
 	"fmt"
-	"log"
 	"os"
 	"strconv"
 	"testing"
@@ -19,60 +18,6 @@ import (
 const (
 	metalDedicatedConnIDEnvVar = "TF_ACC_METAL_DEDICATED_CONNECTION_ID"
 )
-
-func init() {
-	resource.AddTestSweepers("equinix_metal_vrf", &resource.Sweeper{
-		Name: "equinix_metal_vrf",
-		Dependencies: []string{
-			"equinix_metal_device",
-			"equinix_metal_virtual_circuit",
-			// TODO: add sweeper when offered
-			// "equinix_metal_reserved_ip_block",
-		},
-		F: testSweepVRFs,
-	})
-}
-
-func testSweepVRFs(region string) error {
-	log.Printf("[DEBUG] Sweeping VRFs")
-	config, err := acceptance.GetConfigForNonStandardMetalTest()
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting configuration for sweeping VRFs: %s", err)
-	}
-	metal := config.NewMetalClient()
-	ps, _, err := metal.Projects.List(nil)
-	if err != nil {
-		return fmt.Errorf("[INFO][SWEEPER_LOG] Error getting project list for sweeping VRFs: %s", err)
-	}
-	pids := []string{}
-	for _, p := range ps {
-		if acceptance.IsSweepableTestResource(p.Name) {
-			pids = append(pids, p.ID)
-		}
-	}
-	dids := []string{}
-	for _, pid := range pids {
-		ds, _, err := metal.VRFs.List(pid, nil)
-		if err != nil {
-			log.Printf("Error listing VRFs to sweep: %s", err)
-			continue
-		}
-		for _, d := range ds {
-			if acceptance.IsSweepableTestResource(d.Name) {
-				dids = append(dids, d.ID)
-			}
-		}
-	}
-
-	for _, did := range dids {
-		log.Printf("Removing VRFs %s", did)
-		_, err := metal.VRFs.Delete(did)
-		if err != nil {
-			return fmt.Errorf("Error deleting VRFs %s", err)
-		}
-	}
-	return nil
-}
 
 func TestAccMetalVRF_basic(t *testing.T) {
 	var vrf metalv1.Vrf
